react-wavesurfer: Introduce prerender prop and render audioPeaks if set
We can prerender the audio peaks without loading the track by adding a
prerender property to the ReactWavesurfer component and then detecting
it and only rendering the audioPeaks without actually loading the audio.
This is extremely important to avoiding preloading a large number of
tracks at the time of creation.